requirements.txt
psycopg2-binary==2.9.5
kafka-python==2.0.2
pydantic==1.10.2
Flask==2.1.0
Flask-WTF==1.0.1
prometheus-client==0.15.0
sqlalchemy==1.4.44
flask-cors==3.0.10